Check add-on website link destination and redirect instead of clicking the link (#967)
Clicking this link causes a redirect, and with the latest version of Firefox this means we return early from the click. We could wait for the target URL, however it is not good practice to visit third party websites. Instead, as we have hard-coded the add-on, I have simply hard-coded the expected URL and check that we are redirected here when a request is made. Ideally we would generate test data and have full control of the environment, but this is not currently possible.