Avoid maximising the window whilst running the tests unless necessary (#972)
Due to a bug in Marionette, maximising causes an issue when the window is already maximised. I do not see a reason for maximising the window, and the tests are passing without this, so in order to migrate this suite to run against GeckoDriver/Marionette I have removed the calls to maximize_window. See https://bugzilla.mozilla.org/show_bug.cgi?id=1375185 for more details.